// src/generated/prisma/models/Constancia.ts
export type ConstanciaWhereInput = {
  AND?: Prisma.ConstanciaWhereInput | Prisma.ConstanciaWhereInput[]
  OR?: Prisma.ConstanciaWhereInput[]
  NOT?: Prisma.ConstanciaWhereInput | Prisma.ConstanciaWhereInput[]
  id?: Prisma.StringFilter<"Constancia"> | string
  folio?: Prisma.StringFilter<"Constancia"> | string
  validationHash?: Prisma.StringFilter<"Constancia"> | string
  recipientName?: Prisma.StringFilter<"Constancia"> | string
  courseName?: Prisma.StringFilter<"Constancia"> | string
  standardCode?: Prisma.StringNullableFilter<"Constancia"> | string | null
  issuedAt?: Prisma.DateTimeFilter<"Constancia"> | Date | string
  status?: Prisma.EnumConstanciaStatusFilter<"Constancia"> | $Enums.ConstanciaStatus
  createdAt?: Prisma.DateTimeFilter<"Constancia"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"Constancia"> | Date | string
}